🔧 (#305) 강의 상태 관련 타입을 통합하여 코드 정리 및 개선
