+++ frontend/store/useLectureStatusStore.ts
@@ -1,19 +1,15 @@
+import { StudentLectureStatus } from "@/types/lectures/fetchStudentLectureDetailTypes";
 import { create } from "zustand";
 
-export type LectureStatus =
-  | "beforeLecture"
-  | "onLecture"
-  | "makeQuiz"
-  | "checkDashboard";
-
 interface LectureStatusStore {
-  lectureStatus: LectureStatus | null;
-  setLectureStatus: (status: LectureStatus) => void;
+  lectureStatus: StudentLectureStatus | null;
+  setLectureStatus: (status: StudentLectureStatus) => void;
   clearLectureStatus: () => void;
 }
 
 export const useLectureStatusStore = create<LectureStatusStore>((set) => ({
   lectureStatus: null,
-  setLectureStatus: (status: LectureStatus) => set({ lectureStatus: status }),
+  setLectureStatus: (status: StudentLectureStatus) =>
+    set({ lectureStatus: status }),
   clearLectureStatus: () => set({ lectureStatus: null }),
 }));
